+++ package.json
@@ -18,6 +18,7 @@
 		"@radix-ui/react-dropdown-menu": "^2.1.1",
 		"@radix-ui/react-icons": "^1.3.0",
 		"@radix-ui/react-label": "^2.1.0",
+		"@radix-ui/react-separator": "^1.1.0",
 		"@radix-ui/react-slot": "^1.1.0",
 		"@radix-ui/react-toast": "^1.2.1",
 		"class-variance-authority": "^0.7.0",
