+++ src/auth.ts
@@ -22,102 +22,58 @@ export async function getUser(email: string) {
 }
 
 export const { auth, signIn, signOut, handlers } = NextAuth({
-	adapter: PrismaAdapter(prisma) as Adapter,
-	trustHost: true,
-	pages: {
-		signIn: "/signin",
-		newUser: "/signup",
-	},
-	secret: process.env.AUTH_SECRET,
-	session: { strategy: "jwt" },
-	callbacks: {
-		async authorized({ auth, request: { nextUrl } }) {
-			const isLoggedIn = !!auth?.user;
-			const isOnLoginPage = nextUrl.pathname.startsWith("/signin");
-			const isOnSignupPage = nextUrl.pathname.startsWith("/signup");
+	// adapter: PrismaAdapter(prisma) as Adapter,
 
-			if (isLoggedIn) {
-				if (isOnLoginPage || isOnSignupPage) {
-					return Response.redirect(new URL("/", nextUrl));
-				}
-			}
-			return true;
-		},
-		async jwt({ token, user }) {
-			if (user) {
-				token = { ...token, id: user.id, role: user.role };
-			}
-
-			return token;
-		},
-		async session({ session, token }) {
-			if (token) {
-				const { id, role } = token as { id: string; role: string };
-				const { user } = session;
-
-				session = { ...session, user: { ...user, id, role } };
-			}
-
-			return session;
-		},
-	},
-	// ...authConfig,
-	providers: [
-		Credentials({
-			// You can specify which fields should be submitted, by adding keys to the `credentials` object.
-			// e.g. domain, username, password, 2FA token, etc.
-			credentials: {
-				Email: {},
-				Password: {},
-			},
-			async authorize(credentials) {
-				const parsedCredentials = signInSchema.safeParse(credentials);
-
-				if (parsedCredentials.success) {
-					const { email, password } = parsedCredentials.data;
-					const user = await getUser(email);
-
-					if (!user) return null;
-
-					const encoder = new TextEncoder();
-					const saltedPassword = encoder.encode(password + user.salt);
-					const hashedPasswordBuffer = await crypto.subtle.digest(
-						"SHA-512",
-						saltedPassword,
-					);
-					const hashedPassword = getStringFromBuffer(hashedPasswordBuffer);
-
-					if (hashedPassword === user.password) {
-						return user;
-					}
-					return null;
-				}
-
-				return null;
-			},
-		}),
-		Resend({
-			from: "no-reply@send.ningxikeji.com",
-		}),
-
-		//NOTE:如果要启用 Nodemailer，不能用 Middleware 推荐的 auth.config.ts 方式配置，否则会抛出如下错误：
-		// Error: The edge runtime does not support Node.js 'stream' module.
-		// see https://github.com/nextauthjs/next-auth/issues/10919 。
-
-		Nodemailer({
-			server: {
-				host: process.env.EMAIL_SERVER_HOST,
-				port: process.env.EMAIL_SERVER_PORT,
-				auth: {
-					user: process.env.EMAIL_SERVER_USER,
-					pass: process.env.EMAIL_SERVER_PASSWORD,
-				},
-			},
-			from: process.env.EMAIL_FROM,
-		}),
-		GitHub,
-		Google,
-	],
+	...authConfig,
+	// providers: [
+	// 	Credentials({
+	// 		// You can specify which fields should be submitted, by adding keys to the `credentials` object.
+	// 		// e.g. domain, username, password, 2FA token, etc.
+	// 		credentials: {
+	// 			Email: {},
+	// 			Password: {},
+	// 		},
+	// 		async authorize(credentials) {
+	// 			const parsedCredentials = signInSchema.safeParse(credentials);
+	// 			if (parsedCredentials.success) {
+	// 				const { email, password } = parsedCredentials.data;
+	// 				const user = await getUser(email);
+	// 				if (!user) return null;
+	// 				const encoder = new TextEncoder();
+	// 				const saltedPassword = encoder.encode(password + user.salt);
+	// 				const hashedPasswordBuffer = await crypto.subtle.digest(
+	// 					"SHA-512",
+	// 					saltedPassword,
+	// 				);
+	// 				const hashedPassword = getStringFromBuffer(hashedPasswordBuffer);
+	// 				if (hashedPassword === user.password) {
+	// 					return user;
+	// 				}
+	// 				return null;
+	// 			}
+	// 			return null;
+	// 		},
+	// 	}),
+	// 	Resend({
+	// 		from: "no-reply@send.ningxikeji.com",
+	// 	}),
+	// 	//NOTE:如果要启用 Nodemailer，不能用 Middleware 推荐的 auth.config.ts 方式配置，否则会抛出如下错误：
+	// 	// Error: The edge runtime does not support Node.js 'stream' module.
+	// 	// see https://github.com/nextauthjs/next-auth/issues/10919 。
+	// 	Nodemailer({
+	// 		server: {
+	// 			host: process.env.EMAIL_SERVER_HOST,
+	// 			port: process.env.EMAIL_SERVER_PORT,
+	// 			auth: {
+	// 				user: process.env.EMAIL_SERVER_USER,
+	// 				pass: process.env.EMAIL_SERVER_PASSWORD,
+	// 			},
+	// 		},
+	// 		from: process.env.EMAIL_FROM,
+	// 	}),
+	// 	GitHub({ allowDangerousEmailAccountLinking: true }),
+	// 	Google,
+	// ],
 	logger: {
 		error(code, ...message) {
 			console.error(code, message);
