+++ src/middleware.ts
@@ -4,17 +4,17 @@ import NextAuth from "next-auth";
 const { auth } = NextAuth(authConfig);
 
 export default auth(async function middleware(req) {
-	const publicPaths = ["/login", "/signup", "/api/auth", "/ex"];
-	const isPublicRoute =
-		publicPaths.some((path) => req.nextUrl.pathname.startsWith(path)) ||
-		req.nextUrl.pathname === "/";
-	if (!req.auth && !isPublicRoute) {
-		return Response.redirect(new URL("/", req.nextUrl.origin));
-	}
+	// const publicPaths = ["/login", "/signup", "/api/auth", "/ex"];
+	// const isPublicRoute =
+	// 	publicPaths.some((path) => req.nextUrl.pathname.startsWith(path)) ||
+	// 	req.nextUrl.pathname === "/";
+	// if (!req.auth && !isPublicRoute) {
+	// 	return Response.redirect(new URL("/", req.nextUrl.origin));
+	// }
 });
 
 // import { auth } from "@/auth";
-
+// export const { auth: middleware } = NextAuth(authConfig);
 // export default auth((req) => {
 // 	const publicPaths = ["/login", "/signup", "/api/auth", "/ex"];
 // 	const isPublicRoute =
